update: rename detector to tracker
